code/traffic_rule_aware_mp/requirements.txt
numpy>=1.22
pandas>=1.4
pyyaml>=6.0
matplotlib>=3.7
shapely>=2.0
tqdm>=4.64

# --- Dataset-specific (install only what you need) ---
# nuScenes
nuscenes-devkit>=1.1.10

# Argoverse 2 (Sensor)
av2>=0.2.2

# Waymo Perception (with maps for v1.4.2/1.4.3)
# TensorFlow build must match the wheel name; adjust for your platform/Python.
tensorflow==2.11.0
waymo-open-dataset-tf-2-11-0; platform_system != "Windows"
